refactor(bot): introduce platform abstraction
Extract platform-specific bot logic (Slack, GitHub) into a `platforms/`
registry, replacing scattered `switch (platform)` blocks and the
previous `withBotPlatformAuthContext`, `getPlatformIdentity`,
`getRequesterInfo`, `getPlatformContext`, and `isGitHubBotEnabled`
helpers with a `BotPlatform` interface implemented per-platform. Each
platform now owns its identity extraction, link-account prompt,
conversation context, requester info, webhook handling, and chat-event
hooks.
