+++ apps/web/src/app/api/chat/link-account/route.ts
@@ -13,11 +13,10 @@ import {
   canKiloUserAccessPlatformIntegration,
   getPlatformIntegration,
 } from '@/lib/bot/platform-helpers';
+import { botPlatforms } from '@/lib/bot/platforms';
 import { processLinkedMessage } from '@/lib/bot/run';
-import { withBotPlatformAuthContext } from '@/lib/bot/platform-auth-context';
 import { Message, ThreadImpl, type Thread } from 'chat';
 import type { User } from '@kilocode/db';
-import { PLATFORM } from '@/lib/integrations/core/constants';
 
 function errorPage(title: string, message: string, status: number): Response {
   return new Response(
@@ -91,10 +90,10 @@ export async function GET(request: Request) {
 
   const { contextKey, identity, thread, message } = linkPayload;
 
-  if (identity.platform === PLATFORM.GITHUB) {
+  if (!botPlatforms.require(identity.platform).usesGenericLinkAccountRoute) {
     return errorPage(
       'Link Not Supported',
-      'GitHub account links must be created from the GitHub link page.',
+      `${identity.platform} account links must be created from the platform-specific link page.`,
       400
     );
   }
@@ -145,13 +144,16 @@ async function reprocessLinkedMessage(
     const platformIntegration = await getPlatformIntegration(identity);
     if (!platformIntegration) return;
 
-    await withBotPlatformAuthContext(platformIntegration, async () => {
-      await processLinkedMessage({
-        thread,
-        message,
-        platformIntegration,
-        user,
-      });
+    await botPlatforms.require(platformIntegration.platform).withAuthContext({
+      platformIntegration,
+      fn: async () => {
+        await processLinkedMessage({
+          thread,
+          message,
+          platformIntegration,
+          user,
+        });
+      },
     });
   } catch (error) {
     console.error('[Bot] Failed to reprocess linked message:', error);
