+++ apps/web/src/app/api/integrations/github/callback/route.ts
@@ -26,7 +26,7 @@ import { linkKiloUser } from '@/lib/bot-identity';
 import { bot } from '@/lib/bot';
 import { isOrganizationMember } from '@/lib/organizations/organizations';
 import { PLATFORM } from '@/lib/integrations/core/constants';
-import { isGitHubBotEnabled } from '@/lib/bot/platform-helpers';
+import { botPlatforms } from '@/lib/bot/platforms';
 
 function htmlPage(title: string, message: string, status = 200): Response {
   return new Response(
@@ -69,7 +69,7 @@ async function handleGitHubBotLinkCallback(request: NextRequest, user: { id: str
     return htmlPage('Link Failed', 'No matching GitHub integration was found.', 404);
   }
 
-  if (!isGitHubBotEnabled(integration)) {
+  if (!botPlatforms.require(PLATFORM.GITHUB).isEnabledForBot(integration)) {
     return htmlPage('Link Unavailable', 'GitHub linking is not enabled for this integration.', 404);
   }
