+++ apps/web/src/app/github/link/route.ts
@@ -5,8 +5,10 @@ import { APP_URL } from '@/lib/constants';
 import { createGitHubBotLinkState } from '@/lib/bot/github-link-state';
 import { verifyGitHubLinkToken } from '@/lib/bot/github-link-token';
 import { getGitHubAppCredentials } from '@/lib/integrations/platforms/github/app-selector';
-import { getPlatformIntegrationById, isGitHubBotEnabled } from '@/lib/bot/platform-helpers';
+import { getPlatformIntegrationById } from '@/lib/bot/platform-helpers';
+import { botPlatforms } from '@/lib/bot/platforms';
 import { isOrganizationMember } from '@/lib/organizations/organizations';
+import { PLATFORM } from '@/lib/integrations/core/constants';
 
 const GITHUB_AUTHORIZE_URL = 'https://github.com/login/oauth/authorize';
 const GITHUB_CALLBACK_PATH = '/api/integrations/github/callback';
@@ -53,7 +55,7 @@ export async function GET(request: NextRequest) {
     return errorPage('Link Failed', 'No matching GitHub integration was found.', 404);
   }
 
-  if (!isGitHubBotEnabled(integration)) {
+  if (!botPlatforms.require(PLATFORM.GITHUB).isEnabledForBot(integration)) {
     return errorPage(
       'Link Unavailable',
       'GitHub linking is not enabled for this integration.',
