+++ apps/web/src/lib/bot/platform-helpers.test.ts
@@ -20,24 +20,18 @@ jest.mock('@/lib/organizations/organizations', () => ({
 import { PLATFORM } from '@/lib/integrations/core/constants';
 import {
   canKiloUserAccessPlatformIntegration,
-  getBotDocumentationUrl,
   getGitHubRepositoryReference,
-  getPlatformIdentity,
   getPlatformIntegration,
   getPlatformIntegrationByBotUserId,
-  isGitHubRepositoryLinked,
   getPlatformIntegrationById,
-  isGitHubBotEnabled,
+  isGitHubRepositoryLinked,
 } from './platform-helpers';
 import type { PlatformIntegration } from '@kilocode/db';
 import type { Thread, Message } from 'chat';
 
-const mockGetInstallationId = jest.fn();
-
 describe('platform helpers', () => {
   beforeEach(() => {
     mockLimit.mockReset();
-    mockGetInstallationId.mockReset();
     mockIsOrganizationMember.mockReset();
   });
 
@@ -109,76 +103,6 @@ describe('platform helpers', () => {
 
     expect(result).toBeNull();
     expect(mockLimit).not.toHaveBeenCalled();
-  });
-
-  it('extracts GitHub identity from chat adapter messages', async () => {
-    const message = {
-      author: { userId: '12345' },
-      raw: {
-        type: 'issue_comment',
-      },
-    };
-    mockGetInstallationId.mockResolvedValue(98765);
-
-    const identity = await getPlatformIdentity(
-      { adapter: { name: PLATFORM.GITHUB }, id: 'github:acme/widgets:42' } as Thread,
-      message as Message,
-      mockGetInstallationId
-    );
-
-    expect(mockGetInstallationId).toHaveBeenCalledWith({
-      adapter: { name: PLATFORM.GITHUB },
-      id: 'github:acme/widgets:42',
-    });
-    expect(identity).toEqual({
-      platform: PLATFORM.GITHUB,
-      teamId: '98765',
-      userId: '12345',
-    });
-  });
-
-  it('throws when the GitHub adapter cannot resolve the installation id', async () => {
-    const message = {
-      author: { userId: '12345' },
-      raw: {
-        type: 'issue_comment',
-      },
-    } as Message;
-    mockGetInstallationId.mockResolvedValue(null);
-
-    await expect(
-      getPlatformIdentity(
-        { adapter: { name: PLATFORM.GITHUB }, id: 'github:acme/widgets:42' } as Thread,
-        message,
-        mockGetInstallationId
-      )
-    ).rejects.toThrow('Could not find GitHub installation ID for thread github:acme/widgets:42');
-  });
-
-  describe('isGitHubBotEnabled', () => {
-    function integrationWithMetadata(
-      metadata: PlatformIntegration['metadata']
-    ): PlatformIntegration {
-      return { metadata } as PlatformIntegration;
-    }
-
-    it('returns true only when metadata.bot_enabled is the boolean true', () => {
-      expect(isGitHubBotEnabled(integrationWithMetadata({ bot_enabled: true }))).toBe(true);
-    });
-
-    it('returns false when metadata is missing the flag', () => {
-      expect(isGitHubBotEnabled(integrationWithMetadata({}))).toBe(false);
-      expect(isGitHubBotEnabled(integrationWithMetadata(null))).toBe(false);
-    });
-
-    it('returns false for truthy non-boolean values to avoid accidental enables', () => {
-      expect(isGitHubBotEnabled(integrationWithMetadata({ bot_enabled: 'true' }))).toBe(false);
-      expect(isGitHubBotEnabled(integrationWithMetadata({ bot_enabled: 1 }))).toBe(false);
-    });
-
-    it('returns false when explicitly disabled', () => {
-      expect(isGitHubBotEnabled(integrationWithMetadata({ bot_enabled: false }))).toBe(false);
-    });
   });
 
   describe('getGitHubRepositoryReference', () => {
@@ -304,17 +228,5 @@ describe('platform helpers', () => {
         canKiloUserAccessPlatformIntegration({} as PlatformIntegration, 'user-1')
       ).resolves.toBe(false);
     });
-  });
-
-  it('returns platform-specific bot documentation URLs', () => {
-    expect(getBotDocumentationUrl(PLATFORM.SLACK)).toBe(
-      'https://kilo.ai/docs/code-with-ai/platforms/slack'
-    );
-    expect(getBotDocumentationUrl(PLATFORM.GITHUB)).toBe(
-      'https://kilo.ai/docs/code-with-ai/platforms/slack'
-    );
-    expect(getBotDocumentationUrl(PLATFORM.DISCORD)).toBe(
-      'https://kilo.ai/docs/code-with-ai/platforms/slack'
-    );
   });
 });
