+++ apps/web/src/lib/bot/platform-helpers.ts
@@ -4,11 +4,7 @@ import { eq, and, sql } from 'drizzle-orm';
 import { platform_integrations, type PlatformIntegration } from '@kilocode/db';
 import type { Message, Thread } from 'chat';
 import type { GitHubRawMessage } from '@chat-adapter/github';
-import { PLATFORM } from '@/lib/integrations/core/constants';
-import { type SlackEvent } from '@chat-adapter/slack';
 import { isOrganizationMember } from '@/lib/organizations/organizations';
-
-type GetGitHubInstallationId = (thread: Thread) => Promise<string | number | null | undefined>;
 
 export type GitHubRepositoryReference = {
   id: number | null;
@@ -45,48 +41,6 @@ export function getGitHubRepositoryReference(
       parseGitHubRepositoryFullName(thread.id) ??
       parseGitHubRepositoryFullName(thread.channelId),
   };
-}
-
-function getSlackTeamId(message: Message<SlackEvent>): string {
-  const teamId = message.raw.team_id ?? message.raw.team;
-
-  if (!teamId) throw new Error('Expected a teamId in message.raw');
-
-  return teamId;
-}
-
-/**
- * Extract platform identity coordinates from any adapter's message.
- * Extend the switch for Discord / Teams / Google Chat / etc.
- */
-export async function getPlatformIdentity(
-  thread: Thread,
-  message: Message,
-  getGitHubInstallationId: GetGitHubInstallationId
-): Promise<PlatformIdentity> {
-  const platform = thread.adapter.name;
-
-  switch (platform) {
-    case PLATFORM.GITHUB: {
-      const teamId = await getGitHubInstallationId(thread);
-
-      if (!teamId) {
-        throw new Error(`Could not find GitHub installation ID for thread ${thread.id}`);
-      }
-
-      return {
-        platform: PLATFORM.GITHUB,
-        teamId: teamId.toString(),
-        userId: message.author.userId,
-      };
-    }
-    case PLATFORM.SLACK: {
-      const teamId = getSlackTeamId(message as Message<SlackEvent>);
-      return { platform: PLATFORM.SLACK, teamId, userId: message.author.userId };
-    }
-    default:
-      throw new Error(`PlatformNotSupported: ${platform}`);
-  }
 }
 
 /**
@@ -157,17 +111,6 @@ export async function getPlatformIntegrationByBotUserId(
   return integration ?? null;
 }
 
-/**
- * Canary gate for the GitHub bot path. Driven by `metadata.bot_enabled` on the
- * platform integration row so we can enable the bot per-installation without a
- * schema migration. Defaults to false: existing GitHub integrations are not
- * affected until an operator opts them in by setting the flag to true.
- */
-export function isGitHubBotEnabled(integration: PlatformIntegration): boolean {
-  const metadata = (integration.metadata ?? {}) as { bot_enabled?: unknown };
-  return metadata.bot_enabled === true;
-}
-
 export function isGitHubRepositoryLinked(
   integration: PlatformIntegration,
   repository: GitHubRepositoryReference
@@ -186,12 +129,4 @@ export function isGitHubRepositoryLinked(
       linkedRepository.full_name.toLowerCase() === repository.fullName.toLowerCase()
     );
   });
-}
-
-export function getBotDocumentationUrl(platform: string): string {
-  switch (platform) {
-    //TODO(remon): Update when we have specific docs pages for other platforms
-    default:
-      return 'https://kilo.ai/docs/code-with-ai/platforms/slack';
-  }
 }
